Fix failures on btest for NativeClient. * bootstraptest/runner.rb (nacl?): New method to distinguish NaCl   corss build.
* bootstraptest/test_io.rb: Skip unsupported operations.

* bootstraptest/test_literal.rb: ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@38684 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ bootstraptest/test_io.rb
@@ -53,7 +53,7 @@ assert_equal 'ok', %q{
   STDIN.reopen(rw)
   STDIN.reopen(save)
   rw.close
-  File.unlink(tmpname)
+  File.unlink(tmpname) unless RUBY_PLATFORM['nacl']
   :ok
 }
 
@@ -70,7 +70,7 @@ assert_equal 'ok', %q{
   STDIN.print "a"
   STDIN.reopen(save)
   rw.close
-  File.unlink(tmpname)
+  File.unlink(tmpname) unless RUBY_PLATFORM['nacl']
   :ok
 }
